+++ learning_assistant/api.py
@@ -12,7 +12,7 @@ from edx_django_utils.cache import get_cache_key
 from jinja2 import BaseLoader, Environment
 from opaque_keys import InvalidKeyError
 
-from learning_assistant.constants import ACCEPTED_CATEGORY_TYPES, CATEGORY_TYPE_MAP
+from learning_assistant.constants import ACCEPTED_CATEGORY_TYPES, AUDIT_TRIAL_MAX_DAYS, CATEGORY_TYPE_MAP
 from learning_assistant.data import LearningAssistantAuditTrialData, LearningAssistantCourseEnabledData
 from learning_assistant.models import (
     LearningAssistantAuditTrial,
@@ -29,7 +29,6 @@ from learning_assistant.platform_imports import (
     traverse_block_pre_order,
 )
 from learning_assistant.text_utils import html_to_text
-from learning_assistant.utils import get_audit_trial_length_days
 
 log = logging.getLogger(__name__)
 User = get_user_model()
@@ -243,7 +242,7 @@ def get_message_history(courserun_key, user, message_count):
     return message_history
 
 
-def get_audit_trial_expiration_date_from_start_date(start_date, user_id, enrollment_mode):
+def get_audit_trial_expiration_date_from_start_date(start_date):
     """
     Given a start date of an audit trial, return the expiration date of the audit trial.
 
@@ -253,14 +252,12 @@ def get_audit_trial_expiration_date_from_start_date(start_date, user_id, enrollm
 
     Arguments:
     * start_date (datetime): the start date of the audit trial
-    * user_id (int): user id
-    * enrollment_mode (str): enrollment mode of the user
 
     Returns:
     * expiration_date (datetime): the expiration date of the audit trial
     """
-    trial_length_days = get_audit_trial_length_days(user_id, enrollment_mode)
-    expiration_datetime = start_date + timedelta(days=trial_length_days)
+    # Default audit trial length is 14 days
+    expiration_datetime = start_date + timedelta(days=AUDIT_TRIAL_MAX_DAYS)
 
     return expiration_datetime
 
@@ -291,13 +288,12 @@ def get_audit_trial(user):
     )
 
 
-def get_or_create_audit_trial(user, enrollment_mode):
+def get_or_create_audit_trial(user):
     """
     Given a user, return the associated audit trial data, creating a new audit trial for the user if one does not exist.
 
     Arguments:
     * user (User): the user
-    * enrollment_mode (str): enrollment mode of the user
 
     Returns:
     * audit_trial_data (LearningAssistantAuditTrialData): the audit trial data
@@ -306,7 +302,7 @@ def get_or_create_audit_trial(user, enrollment_mode):
         * expiration_date (datetime): the expiration date of the audit trial
     """
     start_date = timezone.now()
-    expiration_date = get_audit_trial_expiration_date_from_start_date(start_date, user.id, enrollment_mode)
+    expiration_date = get_audit_trial_expiration_date_from_start_date(start_date)
 
     audit_trial, _ = LearningAssistantAuditTrial.objects.get_or_create(
         user=user,